+++ widgets/client/messenger/components/bot/Bot.tsx
