+++ widgets/client/messenger/components/bot/Carousel.tsx
