+++ widgets/client/messenger/components/bot/CustomMessage.tsx
